requirements.txt
streamlit==1.40.1
pdfminer.six==20240706
pandas==2.2.2
